Moved metakeys code to store/_metakeys_config
The metakeys code is part of the "domain" logic for
datasource lookup, and so it make sense for it to be with the rest
of this code, which is in `store`, not `objectstore`.
